// src/app/api/model/applied-opportunity-status-history-dto-out.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OpportunityStatus } from './opportunity-status';


export interface AppliedOpportunityStatusHistoryDtoOut { 
    appliedOpportunityId?: number;
    changeReason?: string;
    changedAt?: string;
    changedByFirebaseId?: string;
    changedByUserId?: number;
    changedByUserName?: string;
    id?: number;
    newStatus?: OpportunityStatus;
    notes?: string;
    previousStatus?: OpportunityStatus;
}